Use blue/green switch scripts from release artefact
The script used for the blue/green switch was being taken from
the master branch of the AWS repository. This change ensures
that the version of the script is that included in the release
artefacts.

This ensures that the version of the blue/green switch script
used is consistent with the version of the AWS infrastructure
being used.