MGS-5813 Set catch error for test stages
Specifically in the Mygov pipeline if any of the publishing E2E tests
fail then this would fail the entire pipeline and the webE2E tests would
not run.

This commit adds a catch error for those test stages so that if
a test fails the stage of the pipeline would be set to unstable but the
pipeline would carry on. Overall if this happens then the whole pipeline
will finish with an Unstable result which is what we want.