MGS-5814 Added jobs to build exp/egv
Adds build exp job and egv job and triggered the building of those from
the integration test scheduled job. This means that exp/egv are built
after the tests are ran on int/igv in the morning.

This also adds teardown jobs to the integration testing jobs and removes
them from the end of day teardown