Change end to end job to use the publishing parameter
Change the environment build to pass in 'publishing' instead of 'mygov' to the
endtoend build.
Change the end to end build job  to use the pipeline param instead of site param.
This results in the job id mygov-e2e-publishing instead of mygov-e2e-mygov
Change the test pipeline script to point to the new mygov-e2e-publishing job.